refactor: use messages repo for FindMessagesOfIdentity
