+++ Modules/Messages/src/Messages.Application/Messages/Queries/ListMessages/Handler.cs
@@ -1,12 +1,8 @@
 ﻿using AutoMapper;
-using Backbone.Modules.Messages.Application.Extensions;
 using Backbone.Modules.Messages.Application.Infrastructure.Persistence;
+using Backbone.Modules.Messages.Application.Infrastructure.Persistence.Repository;
 using Backbone.Modules.Messages.Application.Messages.DTOs;
-using Backbone.Modules.Messages.Domain.Entities;
-using Enmeshed.BuildingBlocks.Application.Abstractions.Infrastructure.Persistence.Database;
 using Enmeshed.BuildingBlocks.Application.Abstractions.Infrastructure.UserContext;
-using Enmeshed.BuildingBlocks.Application.Extensions;
-using Enmeshed.DevelopmentKit.Identity.ValueObjects;
 using MediatR;
 
 namespace Backbone.Modules.Messages.Application.Messages.Queries.ListMessages;
@@ -16,19 +12,21 @@ public class Handler : IRequestHandler<ListMessagesQuery, ListMessagesResponse>
     private readonly IMessagesDbContext _dbContext;
     private readonly IMapper _mapper;
     private readonly MessageService _messageService;
+    private readonly IMessagesRepository _messagesRepository;
     private readonly IUserContext _userContext;
 
-    public Handler(IMessagesDbContext dbContext, IUserContext userContext, IMapper mapper, MessageService messageService)
+    public Handler(IMessagesDbContext dbContext, IUserContext userContext, IMapper mapper, MessageService messageService, IMessagesRepository messagesRepository)
     {
         _dbContext = dbContext;
         _userContext = userContext;
         _mapper = mapper;
         _messageService = messageService;
+        _messagesRepository = messagesRepository;
     }
 
     public async Task<ListMessagesResponse> Handle(ListMessagesQuery request, CancellationToken cancellationToken)
     {
-        var dbPaginationResult = await FindMessagesOfIdentity(_userContext.GetAddress(), request);
+        var dbPaginationResult = await _messagesRepository.FindMessagesOfIdentity(_userContext.GetAddress(), request, cancellationToken);
 
         var response = new ListMessagesResponse(_mapper.Map<IEnumerable<MessageDTO>>(dbPaginationResult.ItemsOnPage), request.PaginationFilter, dbPaginationResult.TotalNumberOfItems);
 
@@ -39,26 +37,5 @@ public class Handler : IRequestHandler<ListMessagesQuery, ListMessagesResponse>
         response.PrepareForActiveIdentity(_userContext.GetAddress());
 
         return response;
-    }
-
-    private async Task<DbPaginationResult<Message>> FindMessagesOfIdentity(IdentityAddress identityAddress, ListMessagesQuery request)
-    {
-        var addressOfActiveIdentity = _userContext.GetAddress();
-
-        var query = _dbContext
-            .Set<Message>()
-            .AsQueryable()
-            .IncludeAllReferences();
-
-        if (request.Ids.Any())
-            query = query.WithIdsIn(request.Ids);
-
-        query = query.WithSenderOrRecipient(identityAddress);
-
-        query = query.DoNotSendBeforePropertyIsNotInTheFuture();
-
-        var dbPaginationResult = await query.OrderAndPaginate(d => d.CreatedAt, request.PaginationFilter);
-
-        return dbPaginationResult;
     }
 }
