+++ Modules/Messages/src/Messages.Infrastructure/Persistence/Database/Repository/MessagesRepository.cs
@@ -1,4 +1,4 @@
-using Backbone.Modules.Messages.Application.Infrastructure.Persistence.Repository;
+﻿using Backbone.Modules.Messages.Application.Infrastructure.Persistence.Repository;
 using Backbone.Modules.Messages.Application.Extensions;
 using Backbone.Modules.Messages.Domain.Entities;
 using Backbone.Modules.Messages.Domain.Ids;
@@ -7,6 +7,9 @@ using Enmeshed.BuildingBlocks.Application.Pagination;
 using Microsoft.EntityFrameworkCore;
 using Enmeshed.BuildingBlocks.Application.Extensions;
 using Enmeshed.BuildingBlocks.Application.Abstractions.Infrastructure.UserContext;
+using Backbone.Modules.Messages.Application.Messages.Commands.SendMessage;
+using Enmeshed.DevelopmentKit.Identity.ValueObjects;
+using Backbone.Modules.Messages.Application.Messages.Queries.ListMessages;
 
 namespace Backbone.Modules.Messages.Infrastructure.Persistence.Database.Repository;
 public class MessagesRepository : IMessagesRepository
@@ -29,6 +32,11 @@ public class MessagesRepository : IMessagesRepository
             .FirstWithId(id, cancellationToken);
     }
 
+    async Task<Message> IMessagesRepository.FindPlain(MessageId id, CancellationToken cancellationToken)
+    {
+        return await _readOnlyMessages.FirstWithId(id, cancellationToken);
+    }
+
     async Task<DbPaginationResult<Message>> IMessagesRepository.FindAll(PaginationFilter paginationFilter, CancellationToken cancellationToken)
     {
         return await _readOnlyMessages.OrderAndPaginate(d => d.CreatedAt, paginationFilter);
@@ -46,5 +54,22 @@ public class MessagesRepository : IMessagesRepository
             .FromASpecificSender(sender)
             .WithASpecificRecipientWhoDidNotReceiveTheMessage(recipientDto.Address)
             .CountAsync(cancellationToken);
+    }
+
+    public Task<DbPaginationResult<Message>> FindMessagesOfIdentity(IdentityAddress identityAddress, ListMessagesQuery request, CancellationToken cancellationToken)
+    {
+        var addressOfActiveIdentity = _userContext.GetAddress();
+
+        var query = _readOnlyMessages
+            .AsQueryable()
+            .IncludeAllReferences();
+
+        if (request.Ids.Any())
+            query = query.WithIdsIn(request.Ids);
+
+        query = query.WithSenderOrRecipient(identityAddress);
+        query = query.DoNotSendBeforePropertyIsNotInTheFuture();
+
+        return query.OrderAndPaginate(d => d.CreatedAt, request.PaginationFilter);
     }
 }
